#2 Shop Skeleton: implemented base layout file; moved pages out of the web root
