+++ src/pages/contact-us.php
@@ -1,0 +1,16 @@
+<form method="post" action="/send-contact-us-request">
+    <h2>Contact Us</h2>
+    <div class="field">
+        <label for="contact-name">Your name:</label>
+        <input type="text" id="contact-name" placeholder="John Doe">
+    </div>
+    <div class="field">
+        <label for="contact-email">Your email:</label>
+        <input type="email" id="contact-email" placeholder="email@example.com">
+    </div>
+    <div class="field">
+        <label for="contact-question">Question:</label>
+        <textarea id="contact-question" rows="4" placeholder="Please, enter your question"></textarea>
+    </div>
+    <button type="submit">Submit</button>
+</form>
