+++ src/pages/home.php
@@ -1,0 +1,30 @@
+<!-- @TODO: Implement recently viewed products -->
+<section title="Recently Viewed Products">
+    <h2>Recently Viewed Products</h2>
+    <div class="product-list">
+        <div class="product">
+            <a href="/product-1-url" title="Product 1">
+                <img src="/product-placeholder.png" alt="Product 1" width="200"/>
+            </a>
+            <a href="/product-1-url" title="Product 1">Product 1</a>
+            <span>$33.33</span>
+            <button type="button">Add To Cart</button>
+        </div>
+        <div class="product">
+            <a href="/product-2-url" title="Product 2">
+                <img src="/product-placeholder.png" alt="Product 2" width="200"/>
+            </a>
+            <a href="/product-2-url" title="Product 2">Product 2</a>
+            <span>$66.66</span>
+            <button type="button">Add To Cart</button>
+        </div>
+        <div class="product">
+            <a href="/product-3-url" title="Product 3">
+                <img src="/product-placeholder.png" alt="Product 3" width="200"/>
+            </a>
+            <a href="/product-3-url" title="Product 3">Product 3</a>
+            <span>$99.99</span>
+            <button type="button">Add To Cart</button>
+        </div>
+    </div>
+</section>
