+++ src/pages/product.php
@@ -1,0 +1,7 @@
+<div class="product-page">
+    <img src="/product-placeholder.png" alt="<?= $data['name'] ?>" width="300"/>
+    <h1><?= $data['name'] ?></h1>
+    <p><?= $data['description'] ?></p>
+    <span>$<?= $data['price'] ?></span>
+    <button type="button">Add To Cart</button>
+</div>
