--- web/hello-world.php
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Hello, World!</title>
-</head>
-<body>
-<p style="font-size: 48px;">Hello, World!</p>
-</body>
-</html>
